backend/ai/ai_integration.py
### **`backend/ai/ai_integration.py`** - Vertex AI for move suggestions

from google.cloud import aiplatform
from google.oauth2 import service_account
from config.settings import settings

credentials = service_account.Credentials.from_service_account_file(settings.google_application_credentials)
aiplatform.init(credentials=credentials, project=settings.firebase_project_id)

def suggest_move(game_state):
    model = aiplatform.Model(settings.vertex_ai_model_id)
    response = model.predict([game_state])
    return response[0]['suggested_move']