pyproject.toml
[tool.poetry]
name = "coexistenceboardchain"
version = "0.1.0"
description = "Interactive chess bot promoting coexistence between humans and AI."
authors = ["Beaver"]
license = "GPL-2.0"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
firebase-admin = "^5.2.0"
telethon = "^1.24.0"
python-chess = "^1.999"
google-auth = "^2.6.0"
google-cloud-aiplatform = "^1.12.0"
redis = "^4.3.4"
pydantic = "^1.9.0"
toml = "^0.10.2"
requests = "^2.26.0"
httpx = "^0.23.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest-mock = "^3.14.0"
anyio = "^3.0"
trio = "^0.26.2"
pytest-asyncio = "^0.17"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"